+++ javascript/node/selenium-webdriver/CHANGES.md
@@ -10,6 +10,8 @@ major version release (i.e. 1.0.0).
 
 ### Change Summary
 
+* Added native support for Internet Explorer - the Java Selenium server is no
+    longer required.
 * With the release of [Node 0.12.0](http://blog.nodejs.org/2015/02/06/node-v0-12-0-stable/)
     (finally!), the minimum supported version of Node is now `0.10.x`.
 * The `promise` module is now [Promises/A+](https://promisesaplus.com/)